+++ Realm/Tests/Swift/SwiftArrayTests.swift
@@ -998,8 +998,8 @@ class SwiftArrayTests: RLMTestCase {
         makeEmployee(realm, 40, "C", true)
         try! realm.commitWriteTransaction()
 
-        let sortedByAge = EmployeeObject.allObjectsInRealm(realm).sortedResultsUsingProperty("age", ascending: true)
-        let sortedByName = sortedByAge.sortedResultsUsingProperty("name", ascending: false)
+        let sortedByAge = EmployeeObject.allObjectsInRealm(realm).sortedResultsUsingKeyPath("age", ascending: true)
+        let sortedByName = sortedByAge.sortedResultsUsingKeyPath("name", ascending: false)
 
         XCTAssertEqual(Int32(20), (sortedByAge[0] as! EmployeeObject).age)
         XCTAssertEqual(Int32(40), (sortedByName[0] as! EmployeeObject).age)